requirements.txt
openai>=1.0.0
Pillow>=9.0.0
pyperclip>=1.8.0
requests>=2.25.0